feat: add ROI measurement tooling and tests
